Moved session tests back into source file
